RecuperarContraseña/backend/requirements.txt
flask
flask-cors
flask-mail
bcrypt
python-dotenv
